management/pyproject.toml
[tool.poetry]
name = "management"
version = "0.1.0"
description = "Management Service for AI Recommendation Platform"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.2"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"
drf-spectacular = "^0.27.0"
psycopg2-binary = "^2.9.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.6"
openai = "^1.10.0"
anthropic = "^0.39.0"
gunicorn = "^21.2.0"
python-dotenv = "^1.0.0"
dj-stripe = "^2.9.0"
pinecone = "3.2.2"
langchain-community = "^0.3.20"
langchain-openai = "^0.3.11"
langchain-anthropic = "^0.3.0"
langchain-core = "^0.3.45"
pypdf = "^5.4.0"
pillow = "^11.1.0"
boto3 = "^1.37.29"
numpy = "^1.26.3"
pandas = "^2.2.0"
httpx = "^0.27.0"
pyjwt = "^2.8.0"
beautifulsoup4 = "^4.14.2"
google-genai = "^0.2.2"
aiohttp = "^3.12.15"
redis = "^5.0.0"
psutil = "^7.1.0"
memory-profiler = "^0.61.0"
pyarrow = "^19.0.1"
pydantic = "^2.6.1"
adrf = "^0.1.7"
apify-client = "^1.7.0"
lxml = "^5.0.0"
urllib3 = "^2.0.0"
defusedxml = "^0.7.1"
celery = {extras = ["redis"], version = "^5.3.0"}
flower = "^2.0.0"
django-storages = {extras = ["s3"], version = "^1.14.2"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-django = "^4.7.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest]
testpaths = ["tests"]
DJANGO_SETTINGS_MODULE = "configurations.settings"
